Faster posts

+++ app/v1/views/creations.py
@@ -57,11 +57,12 @@ def rud_creation(creator_id, creation_id):
         return redirect(url_for('app_views.all_creations'))
     if '_method' in request.form.keys() and request.form['_method'] == 'CLEAN':
         posts=creation_obj.posts_no_content#[{'id':i.id, 'title':i.title, 'reference': i.reference} for i in sorted(, key=lambda i:(i.reference, i.fetched_at), reverse=True)]
-        for c in range(1,len(posts)):
-            print(c)
-            if posts[c]['reference'] == posts[c-1]['reference']:
-                print(posts[c]['reference'], posts[c]['title'])
-                storage.get(Post, posts[c]['id']).delete()
+        if posts is not None:
+            for c in range(1,len(posts)):
+                print(c)
+                if posts[c]['reference'] == posts[c-1]['reference']:
+                    print(posts[c]['reference'], posts[c]['title'])
+                    storage.get(Post, posts[c]['id']).delete()
         storage.save()
         return redirect(url_for('app_views.all_creations'))
     if request.method == 'POST':
@@ -77,5 +78,4 @@ def rud_creation(creator_id, creation_id):
     form.creation_creators.data = creation_obj.creator_id
     form.submit.label.text = "Save Changes"
     posts=creation_obj.posts_no_content#[{'id':i.id, 'title':i.title, 'reference': i.reference} for i in sorted(creation_obj.posts_no_content, key=lambda i:i.reference)]
-    print("Len of posts:",len(posts))
     return render_template('user/creation_view.html', creation=creation_obj, posts=posts, form=form)
